requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
PyYAML>=6.0.1
pyarrow>=15.0.0
tqdm>=4.66.0
jupyter>=1.0.0
notebook>=7.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
statsmodels>=0.14.0
matplotlib>=3.8.0
shap>=0.45.0
